requirements.txt
streamlit
google-api-python-client
google-auth
google-auth-httplib2
pandas
numpy
pymatgen
certifi
rich==14.2.0
